dev-4268 made hierarchical filter abstract
